unify generic and simple models
